+++ src/classes/Director.java
@@ -19,7 +19,7 @@ import java.util.logging.Logger;
 
 public class Director extends Thread {
 
-    private int daysRemaining;
+    Semaphore daysRemaining;
     private float totalIncome;
     int rand1;
     int rand2;
@@ -33,7 +33,7 @@ public class Director extends Thread {
     static int descontadoPmB;
     static int descontadoPmN;
 
-    public Director(String studio) {
+    public Director(Semaphore daysRemaining,String studio) {
         this.studio = studio;
         this.daysRemaining = daysRemaining;
         this.totalIncome = 0;
@@ -69,7 +69,7 @@ public class Director extends Thread {
             Bethesda.actualizarJuegosGenerados(Integrador.countJuegosGeneradosB);
             Integrador.countJuegosDLCGeneradosB = 0;
             Bethesda.actualizarJuegosDLCGenerados(Integrador.countJuegosDLCGeneradosB);
-            BethesdaStudio.diasRestantesB = Dashboard.diasParaEntrega;  //Resetea Dias Restantes 
+            daysRemaining.release(BethesdaStudio.diasRestantesB);  //Resetea Dias Restantes 
             Bethesda.estadoDirector.setText("Labores administrativas");
         } else {
             ingresoN += Integrador.countJuegosGeneradosN * 550000;
@@ -78,7 +78,7 @@ public class Director extends Thread {
             Nintendo.actualizarJuegosGenerados(Integrador.countJuegosGeneradosN);
             Integrador.countJuegosDLCGeneradosN = 0;
             Nintendo.actualizarJuegosDLCGenerados(Integrador.countJuegosDLCGeneradosN);
-            NintendoStudio.diasRestantesN =  Dashboard.diasParaEntrega;  //Resetea Dias Restantes
+            daysRemaining.release(NintendoStudio.diasRestantesN);  //Resetea Dias Restantes
             Nintendo.estadoDirector.setText("Labores administrativas");
         }
 
@@ -90,10 +90,10 @@ public class Director extends Thread {
 
             try {
                 if ("B".equals(studio)) {
-                    if (BethesdaStudio.diasRestantesB == 0) {
+                    if (daysRemaining.availablePermits() == 0) {
 
                         Bethesda.actualizarEstadoDirector("Entregando Juegos");
-                        Thread.sleep(1000);  //PASA UN DIA
+                        Thread.sleep(1000*Dashboard.duracionDiasSegundos);  //PASA UN DIA
                         payDayDirector();
                         entregarJuegos();
                     } else {
@@ -101,10 +101,10 @@ public class Director extends Thread {
                         trabajoAdministrativo();
                     }
                 } else {
-                    if (NintendoStudio.diasRestantesN == 0) {
+                    if (daysRemaining.availablePermits() == 0) {
 
                         Nintendo.actualizarEstadoDirector("Entregando Juegos");
-                        Thread.sleep(1000);  //PASA UN DIA
+                        Thread.sleep(1000*Dashboard.duracionDiasSegundos);  //PASA UN DIA
                         payDayDirector();
                         entregarJuegos();
                     } else {
@@ -123,9 +123,9 @@ public class Director extends Thread {
     public void trabajoAdministrativo() {
         try {
             int randomN = new Random().nextInt(1000);
-            Thread.sleep(randomN);
+            Thread.sleep(randomN*Dashboard.duracionDiasSegundos);
             checkPM();
-            Thread.sleep(1000 - randomN);
+            Thread.sleep((1000 - randomN)*Dashboard.duracionDiasSegundos);
             payDayDirector();
 
         } catch (InterruptedException ex) {