+++ src/classes/NintendoStudio.java
@@ -11,6 +11,7 @@ public class NintendoStudio {
     public static int utilidadN = ingresoN - totalPayN;
     public static int juegosGeneradosN = 0;
     public static int juegosDLCGeneradosN = 0;
+    
     // Capacidad de los drive de los desarrolladores
     static int capacidadDriveN = 25;
     static int capacidadDriveNivelN = 20;
@@ -25,6 +26,9 @@ public class NintendoStudio {
     public static Semaphore driveLogicN = new Semaphore(capacidadDriveLogicaN);
     public static Semaphore driveSpritesN = new Semaphore(capacidadDriveSpritesN);
     public static Semaphore driveJuegosN = new Semaphore(Integer.MAX_VALUE);
+    
+    // Semaforo de los dias restantes
+    public static Semaphore diasRestantesSem = new Semaphore(diasRestantesN);
 
     final private boolean active = true;
     public int dayDuration;