+++ src/interfaces/Bethesda.java
@@ -23,12 +23,16 @@ public class Bethesda extends javax.swing.JPanel {
     private int valorSpinnerIntegradoresB;
 
     public Bethesda() {
+        
         initComponents();
-        Bethesda.diasParaEntrega.setText(Integer.toString(BethesdaStudio.diasRestantesB));
-        Director director = new Director("B");
+        Bethesda.diasParaEntrega.setText(Integer.toString(BethesdaStudio.diasRestantesSem.availablePermits()));
+        
+        Director director = new Director(BethesdaStudio.diasRestantesSem,"B");
         director.start();
-        ProjectManager pm = new ProjectManager(10, "B");
+        
+        ProjectManager pm = new ProjectManager(BethesdaStudio.diasRestantesSem, "B");
         pm.start();
+        
         Funciones dia = new Funciones();
         dia.start();
 
@@ -330,7 +334,7 @@ public class Bethesda extends javax.swing.JPanel {
 
         desarrolladoresRestantesB.setFont(new java.awt.Font("Dialog", 1, 14)); // NOI18N
         desarrolladoresRestantesB.setForeground(new java.awt.Color(255, 255, 255));
-        desarrolladoresRestantesB.setText("13");
+        desarrolladoresRestantesB.setText("9");
         add(desarrolladoresRestantesB, new org.netbeans.lib.awtextra.AbsoluteConstraints(210, 130, 20, 20));
 
         jLabel3.setFont(new java.awt.Font("Dialog", 1, 14)); // NOI18N