+++ src/ChangeLog/ChangeLog.md
@@ -1,8 +1,19 @@
 # Registro de Cambios
+## [1.0.35] - 12/10/2023 (Juandb1507)
+
+-  Creación archivo txt
+- Escritura y Lectura del archivo (diasParaEntregar y duracionDiasSegundos)
+- Se agregó a los sleep duracionDiasSegundos
+
+
+## [1.0.34] - 12/10/2023 (Sebasr22)
+
+- Se cambio los dias restantes para la entrega de int a Semaphoro y se ajusto la logica en:
+     BethedaStudio, Bethesda, NintendoStudio, Nintendo, PM y director.
+
 ## [1.0.33] - 12/10/2023 (Juandb1507)
 
 - Corrección al Dashboard Principal
-
 
 ## [1.0.32] - 12/10/2023 (Juandb1507)